Remove icon prop from draggable chip component
