+++ packages/block-editor/src/components/block-draggable/draggable-chip.native.js
@@ -7,12 +7,15 @@ import { View } from 'react-native';
  * WordPress dependencies
  */
 import { dragHandle } from '@wordpress/icons';
+import { useSelect } from '@wordpress/data';
+import { getBlockType } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
  */
 import BlockIcon from '../block-icon';
 import styles from './style.scss';
+import { store as blockEditorStore } from '../../store';
 
 const shadowStyle = {
 	shadowColor: '#000',
@@ -26,11 +29,23 @@ const shadowStyle = {
 	elevation: 5,
 };
 
-export default function BlockDraggableChip( { icon } ) {
+export default function BlockDraggableChip() {
+	const { blockIcon } = useSelect( ( select ) => {
+		const { getBlockName, getDraggedBlockClientIds } = select(
+			blockEditorStore
+		);
+		const draggedBlockClientIds = getDraggedBlockClientIds();
+		const blockName = getBlockName( draggedBlockClientIds[ 0 ] );
+
+		return {
+			blockIcon: getBlockType( blockName )?.icon,
+		};
+	} );
+
 	return (
 		<View style={ [ styles[ 'draggable-chip__container' ], shadowStyle ] }>
 			<BlockIcon icon={ dragHandle } />
-			<BlockIcon icon={ icon } />
+			<BlockIcon icon={ blockIcon } />
 		</View>
 	);
 }
